chore: reorganize PAT routes
