+++ src/routes/token/index.ts
@@ -2,14 +2,7 @@ import { Router } from 'express';
 
 import { asyncWrapper as aw } from '@/utils';
 import { bodyValidator } from '@/validation';
-import {
-  createPATHandler,
-  createPATSchema,
-  tokenHandler,
-  tokenSchema,
-  verifyPATHandler,
-  verifyPATSchema,
-} from './token';
+import { tokenHandler, tokenSchema } from './token';
 import { verifyTokenHandler, verifyTokenSchema } from './verify';
 
 const router = Router();
@@ -38,14 +31,6 @@ router.post(
   '/token/verify',
   bodyValidator(verifyTokenSchema),
   aw(verifyTokenHandler)
-);
-
-router.post('/pat', bodyValidator(createPATSchema), aw(createPATHandler));
-
-router.post(
-  '/pat/verify',
-  bodyValidator(verifyPATSchema),
-  aw(verifyPATHandler)
 );
 
 const tokenRouter = router;
