+++ src/validation/fields.ts
@@ -87,22 +87,24 @@ export const redirectTo = Joi.string()
     }
 
     // * We allow any sub-path of the client url
-    // * With optional hash and query params 
-    if (new RegExp(`^${ENV.AUTH_CLIENT_URL}(\/.*)?([?].*)?([#].*)?$`).test(value)) {
+    // * With optional hash and query params
+    if (
+      new RegExp(`^${ENV.AUTH_CLIENT_URL}(\/.*)?([?].*)?([#].*)?$`).test(value)
+    ) {
       return value;
     }
 
-    // * Check if the value's hostname matches any allowed hostname 
+    // * Check if the value's hostname matches any allowed hostname
     // * Required to avoid shadowing domains
     const hostnames = ENV.AUTH_ACCESS_CONTROL_ALLOWED_REDIRECT_URLS.map(
       (allowed) => {
-        return new URL(allowed).hostname
+        return new URL(allowed).hostname;
       }
     );
 
     const valueUrl = new URL(value);
-    if (!micromatch.isMatch(valueUrl.hostname, hostnames, {nocase: true})) {
-      return helper.error('redirectTo')
+    if (!micromatch.isMatch(valueUrl.hostname, hostnames, { nocase: true })) {
+      return helper.error('redirectTo');
     }
 
     // * We allow any sub-path of the allowed redirect urls.
@@ -128,7 +130,10 @@ export const redirectTo = Joi.string()
     try {
       // * Don't take the query parameters into account
       // * And replace `.` with `/` because of micromatch
-      const urlWithoutParams = `${valueUrl.origin}${valueUrl.pathname}`.replace(/[.]/g, '/');
+      const urlWithoutParams = `${valueUrl.origin}${valueUrl.pathname}`.replace(
+        /[.]/g,
+        '/'
+      );
       const match = micromatch.isMatch(urlWithoutParams, expressions, {
         nocase: true,
       });
@@ -160,6 +165,10 @@ export const refreshToken = uuid
   .description(
     'Refresh token during authentication or when refreshing the JWT'
   );
+
+export const personalAccessToken = uuid
+  .required()
+  .description('Personal access token');
 
 export const token = jwt.optional().description('Access token');
